requirements.txt
Django==3.2.7
asgiref==3.4.1
pytz==2021.1
sqlparse==0.4.2
django-ckeditor
djangorestframework==3.12.4
django-filter==2.4.0
django-extra-fields==3.0.2
